custom update code path for DBCluster
The ModifyDBCluster RDS API call will return an InvalidParameterValue
error if the engine mode for the DBCluster is "serverless" and any of
the following fields have been set on the ModifyDBClusterInput shape:

* PreferredMaintenanceWindow
* PreferredBackupWindow
* Port

This error is returned **even if the value of the field in the
ModifyDBClusterInput shape matches the existing value of the DBCluster
resource**

To get around this issue, I've added a custom update method
implementation for the DBCluster resource that builds the
ModifyDBClusterInput shape and ONLY sets fields on that shape when the
matching Spec field is not nil AND there is a difference detected
between the latest observed value and the desired value for that field.

Fixes Issue aws-controllers-k8s/community#917
