+++ test/e2e/tests/test_db_cluster.py
@@ -110,21 +110,18 @@ class TestDBCluster:
         # We're now going to modify the CopyTagsToSnapshot field of the DB
         # instance, wait some time and verify that the RDS server-side resource
         # shows the new value of the field.
-        # TODO(jaypipes): Uncomment the below test once
-        # https://github.com/aws-controllers-k8s/community/issues/917
-        # has been addressed.
-        #assert dbc_rec['CopyTagsToSnapshot'] == False
-        #updates = {
-        #    "spec": {"copyTagsToSnapshot": True},
-        #}
-        #k8s.patch_custom_resource(ref, updates)
-        #time.sleep(MODIFY_WAIT_AFTER_SECONDS)
+        assert dbc_rec['CopyTagsToSnapshot'] == False
+        updates = {
+            "spec": {"copyTagsToSnapshot": True},
+        }
+        k8s.patch_custom_resource(ref, updates)
+        time.sleep(MODIFY_WAIT_AFTER_SECONDS)
 
-        #aws_res = rds_client.describe_db_clusters(DBClusterIdentifier=db_cluster_id)
-        #assert aws_res is not None
-        #assert len(aws_res['DBClusters']) == 1
-        #dbc_rec = aws_res['DBClusters'][0]
-        #assert dbc_rec['CopyTagsToSnapshot'] == True
+        aws_res = rds_client.describe_db_clusters(DBClusterIdentifier=db_cluster_id)
+        assert aws_res is not None
+        assert len(aws_res['DBClusters']) == 1
+        dbc_rec = aws_res['DBClusters'][0]
+        assert dbc_rec['CopyTagsToSnapshot'] == True
 
         # Delete the k8s resource on teardown of the module
         k8s.delete_custom_resource(ref)
